orchestrator.py
import asyncio
import logging
import os
import yaml
from typing import Dict, List, Any
import zmq.asyncio
from lancedb import connect

# 🛡️ Layer 1 & 2: FileGuard & ShellSandbox (Simplified logic)
class SecurityLayer:
    @staticmethod
    def validate_command(cmd: str):
        blocked = ["rm -rf /", "mkfs"]
        if any(b in cmd for b in blocked):
            raise PermissionError(f"Command blocked by ShellSandbox: {cmd}")
        logging.info(f"ShellSandbox: Command allowed -> {cmd}")

    @staticmethod
    def check_access(path: str):
        allowed_prefixes = ["/apps", "/engines", "/packages", "e:\\IDEAS\\omniclaw"]
        if not any(path.startswith(p) for p in allowed_prefixes):
             # For simulation, we'll just log
             logging.warning(f"FileGuard: Access outside scoped workspace -> {path}")

logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
logger = logging.getLogger("SovereignSentinel")

class SovereignOrchestrator:
    def __init__(self, config_path: str):
        self.config = self._load_config(config_path)
        self.ctx = zmq.asyncio.Context()
        self.kill_switch = asyncio.Event()
        # Vector Memory (LanceDB)
        os.makedirs("./memory", exist_ok=True)
        self.db = connect("./memory/sovereign_vector")
        
    def _load_config(self, path: str) -> Dict:
        with open(path, 'r') as f:
            return yaml.safe_load(f)

    async def start_p2p_mesh(self):
        """ZeroMQ + AES-256-GCM Knowledge Mesh (Boilerplate)"""
        socket = self.ctx.socket(zmq.REP)
        socket.bind("tcp://*:5555")
        logger.info("P2P Knowledge Mesh active on port 5555 [AES-256-GCM]")
        while not self.kill_switch.is_set():
            try:
                msg = await asyncio.wait_for(socket.recv_json(), timeout=1.0)
                # Syncing vector memory or offloading Static Analysis
                logger.info(f"P2P Mesh: Received task from node {msg.get('node_id')}")
                await socket.send_json({"status": "synced", "ack": True})
            except asyncio.TimeoutError:
                continue

    async def cve_to_poc_factory(self, cve_id: str):
        """
        Logic Flow:
        1. Monitor NVD for CVE
        2. StaticSlicer determines reachability via CPG
        3. specialized Worker synthesizes Python exploit
        """
        logger.info(f"🚀 Initializing CVE-to-PoC Factory for {cve_id}")
        
        # Step 1: Reachability Analysis
        reachable = await self.dispatch_task("static-analyst", {"action": "slice", "cve": cve_id})
        if not reachable:
            logger.info(f"CVE {cve_id} is not reachable via public inputs. Aborting.")
            return

        # Step 2: Exploit Synthesis
        logger.info(f"CVE {cve_id} is REACHABLE. Dispatching Exploit Synthesis...")
        exploit = await self.dispatch_task("dynamic-exploiter", {"action": "synthesize", "cve": cve_id})
        
        # Step 3: PoC Validation
        logger.info("Validating synthesized PoC via Playwright Dynamic Agent...")
        validation = await self.dispatch_task("dynamic-exploiter", {"action": "validate", "exploit": exploit})
        
        if validation.get("success"):
            logger.info(f"✅ SUCCESS: Functional PoC generated for {cve_id}")
            self._save_evidence(cve_id, validation)
        else:
            logger.warning(f"❌ FAILED: PoC validation failed for {cve_id}")

    async def dispatch_task(self, worker_id: str, payload: Dict) -> Dict:
        """Simulate dispatching to a specialized worker in the Hive"""
        logger.info(f"Hive: Dispatching to {worker_id}...")
        await asyncio.sleep(2) # Simulating heavy compute
        return {"success": True, "data": "simulated_result"}

    def _save_evidence(self, cve_id: str, data: Any):
        path = f"./evidence/{cve_id}_proof.json"
        os.makedirs("./evidence", exist_ok=True)
        # In a real scenario, this would be encrypted and signed
        logger.info(f"Evidence captured and stored at {path}")

    async def run_forever(self):
        logger.info("Sovereign Sentinel Orchestrator starting...")
        await asyncio.gather(
            self.start_p2p_mesh(),
            self.cve_to_poc_factory("CVE-2026-1337") # Example zero-day trigger
        )

if __name__ == "__main__":
    orchestrator = SovereignOrchestrator("config.pro.yaml")
    try:
        asyncio.run(orchestrator.run_forever())
    except KeyboardInterrupt:
        logger.warning("Kill-switch activated. Shutting down...")
