requirements.txt
# OmniClaw Core Dependencies (v4.1.0)
# Install with: pip install -r requirements.txt

# Async & Networking
asyncio>=3.4.3
aiohttp>=3.9.0
websockets>=12.0
httpx>=0.25.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# AI/ML APIs
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0

# Vector Database
faiss-cpu>=1.7.4
# Or use faiss-gpu for CUDA support

# Embeddings
sentence-transformers>=2.2.0

# Messaging
python-telegram-bot>=20.0
discord.py>=2.3.0
slack-sdk>=3.19.0
matrix-nio>=0.23.0
# whatsapp-web.js requires Node.js

# System Monitoring
psutil>=5.9.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic-settings>=2.0
tenacity>=8.0.0

# Utilities
tqdm>=4.66.0
rich>=13.0.0
colorama>=0.4.6

# Security & Encryption
cryptography>=41.0.0
pydantic>=2.0.0
pycryptodome>=3.19.0        # AES-256-GCM for hive_sync P2P

# P2P Networking
pyzmq>=25.0.0               # ZeroMQ for hive_sync mesh

# Scheduler
croniter>=1.3.0            # Cron expression parsing for scheduler

# Optional: For local LLMs
# llama-cpp-python>=0.2.0
# transformers>=4.35.0
# torch>=2.1.0

# Voice & Audio
SpeechRecognition>=3.10.0
PyAudio>=0.2.13

# Advanced Features
GitPython>=3.1
watchdog>=3.0

# MCP (Model Context Protocol)
mcp>=1.1.2
fastmcp>=0.4.1

# LLM Routing & Council
litellm>=1.0.0

# Quantum Gateway
qiskit>=1.0.0
qiskit_ibm_runtime>=0.20.0

# Recommendation Engine
chromadb>=0.4.20

# Browser Automation
browser-use>=0.1.0
playwright>=1.40.0
langchain_openai>=0.0.3

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.7.0

# GUI
nicegui>=1.4.0

# Advanced Scraping & Decensoring
heretic-llm>=0.1.0
scrapling[all]>=0.1.0

# Observability
langwatch>=0.1.0
